requirements.txt
# RSA Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1          # The web framework we use
uvicorn[standard]==0.24.0 # ASGI server to run FastAPI

# Database
sqlalchemy==2.0.23        # ORM for database operations

# Data Validation
pydantic==2.5.0           # Data validation (used by FastAPI)

# Authentication
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4    # Password hashing (if needed later)

# Utilities
python-dotenv==1.0.0      # Load settings from .env file
python-multipart==0.0.6   # Handle form data uploads

